revert change to support Decimal for config params since they are not supported as a config param type

+++ homeassistant/components/ozw/websocket_api.py
@@ -168,7 +168,7 @@ def websocket_get_config_parameters(hass, connection, msg):
         vol.Required(NODE_ID): vol.Coerce(int),
         vol.Optional(OZW_INSTANCE, default=1): vol.Coerce(int),
         vol.Required(PARAMETER): vol.Coerce(int),
-        vol.Required(VALUE): vol.Any(bool, vol.Coerce(float), str),
+        vol.Required(VALUE): vol.Any(bool, vol.Coerce(int), str),
     }
 )
 def websocket_set_config_parameter(hass, connection, msg):
